server: use CUDA frame as NVENC encoder source
Previously, a CPU-based frame was used for the copy of the vulkan output
frame. This meant a copy from GPU to CPU and back for the NVENC encode.
The new version copies the vulkan output frame to a GPU-based
CUDA-backed frame, which is then passed to the encoder on the GPU side.

(I'm not the author of this)